+++ timely/src/dataflow/channels/pushers/buffer.rs
@@ -96,6 +96,7 @@ impl<T, C: Container, P: Push<BundleCore<T, C>>> BufferCore<T, C, P> where T: Eq
 impl<T, D: Data, P: Push<BundleCore<T, Vec<D>>>> BufferCore<T, Vec<D>, P> where T: Eq+Clone {
     fn ensure_buffer(&mut self) -> &mut Vec<D> {
         if self.buffer.is_none() {
+            // TODO: Vec::with_capacity()
             self.buffer = Some(RefOrMut::Ref(&Vec::new()).assemble(&mut self.allocation));
         }
         self.buffer.as_mut().unwrap()
